ai-agent/requirements.txt
google-adk>=0.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
toolbox-core>=0.1.0
